dockerapp: Introduce wrapper class for package manager
The original implementation of this functionality is based on the
<=0.8 version of Docker App. The 1.0 version of Docker App includes
some new and improved ways of doing things. Namely - "bundles" will now
be stored in the Docker Registry (pinned with sha256 hashes) so that we
no longer have to upload .dockerapp files into the Tuf reposerver.

The 1.0 implementation of this package manager is very very similar to
the current, but they are different enough that trying to keep them
in a single package manager implementation is difficult to maintain.
Additionally, we should be able to eventually remove this current
implementation once we work through a migration/deprecation period.

With that in mind - this change turns the DockerAppManager class into a
simple wrapper that calls the "DockerAppStandalone" functions. This sets
things up so we can create a new "DockerAppBundles" implementation that
can be detected at runtime and used.
